Moraleja: usar int para los hilos

+++ tiny_mc/tiny_mc.cu
@@ -10,9 +10,9 @@
 
 __global__ void simulate_kernel(float* __restrict__ heats, float* __restrict__ heats_squared)
 {
-    unsigned long gtid = blockIdx.x * blockDim.x + threadIdx.x;
+    int gtid = blockIdx.x * blockDim.x + threadIdx.x;
     if (gtid >= PHOTONS) return;
-    unsigned long btid = threadIdx.x;
+    int btid = threadIdx.x;
 
     // Fase 1: Incialización
     __shared__ float heats_local[SHELLS];
